+++ v2/examples/lists/restore_list.go
@@ -1,0 +1,13 @@
+package examples
+
+import (
+	"context"
+
+	"github.com/trycourier/courier-go/v2"
+)
+
+func restoreList() {
+	client := courier.CreateClient("<YOUR_AUTH_TOKEN>", nil)
+
+	client.RestoreList(context.Background(), "my-list")
+}