+++ v2/shared.go
@@ -1,0 +1,6 @@
+package courier
+
+type PagingResponse struct {
+	Cursor *string
+	More   bool
+}