+++ v2/courier.go
@@ -1,6 +1,6 @@
 package courier
 
-const version = "2.1.0"
+const version = "2.2.0"
 
 // Client lets you communicate with the Courier API
 type Client struct {