+++ v2/examples/brands/put_brand.go
@@ -1,0 +1,15 @@
+package examples
+
+import (
+	"context"
+
+	"github.com/trycourier/courier-go/v2"
+)
+
+func putBrand() {
+	client := courier.CreateClient("<YOUR_AUTH_TOKEN>", nil)
+
+	client.PutBrand(context.Background(), "my-brand", courier.PutBrandBody{
+		Name: "my-updated-brand",
+	})
+}